Put implementation of each tab in a separated file
