+++ portal/src/screens/api-resources/APIResourceDetailsScreenApplicationsTab.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function APIResourceDetailsScreenApplicationsTab(): JSX.Element {
+  return <div>TODO</div>;
+}
