+++ portal/src/screens/api-resources/APIResourceDetailsScreenTestTab.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function APIResourceDetailsScreenTestTab(): JSX.Element {
+  return <div>TODO</div>;
+}
